Minor Refactoring & Reorganization
